requirements.txt
 numpy==1.19.4
 azure-storage-blob==12.6.0
 azure-identity==1.5.0
 environs==9.2.0
 flake8==3.8.4
 pylint==2.6.0
 pytest==6.1.2
 autopep8==1.5.4